Add support for multiple commanders in the match details. Reorganized match details components and types
