+++ src/components/matchHistory/types/MatchDisplayCommander.ts
@@ -1,0 +1,4 @@
+export type MatchDisplayCommander = {
+    name: string;
+    id: string | undefined;
+};
