+++ src/components/matchHistory/types/MatchDisplayPlayer.ts
@@ -1,0 +1,7 @@
+import { MatchDisplayCommander } from "./MatchDisplayCommander";
+
+export type MatchDisplayPlayer = {
+    name: string;
+    commanders: MatchDisplayCommander[];
+    isWinner: boolean;
+};
